Moved all functions to main_globus_archiver.py so GlobusArchiver.py can successfully report that the incorrect version of python is being used with getting SyntaxError
